FirebaseScripts/requirements.txt
pyrebase4==4.4.3
firebase=3.0.1